move plugin commands to init file of plugin
