+++ electrum/plugins/labels/__init__.py
@@ -1,4 +1,11 @@
 from electrum.i18n import _
+from electrum.commands import plugin_command
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from .labels import LabelsPlugin
+    from electrum.commands import Commands
+
 
 fullname = _('LabelSync')
 description = ' '.join([
@@ -7,3 +14,15 @@ description = ' '.join([
 ])
 available_for = ['qt', 'qml', 'cmdline']
 
+plugin_name = "labels"
+
+@plugin_command('w', plugin_name)
+async def push(self: 'Commands', plugin: 'LabelsPlugin' = None, wallet=None) -> int:
+    """ push labels to server """
+    return await plugin.push_thread(wallet)
+
+
+@plugin_command('w', plugin_name)
+async def pull(self: 'Commands', plugin: 'LabelsPlugin' = None, wallet=None, force=False) -> int:
+    """ pull labels from server """
+    return await plugin.pull_thread(wallet, force=force)
